refactor(jsruntime): refactoring
