+++ libs/jsruntime/Cargo.toml
@@ -23,7 +23,6 @@ jsparser = { path = "../jsparser", features = ["location"] }
 logging = { path = "../logging" }
 paste = "1.0.15"
 rustc-hash = "2.1.2"
-static_assertions = "1.1.0"
 thiserror = "2.0.18"
 
 [dev-dependencies]
