+++ libs/jsruntime/src/types/mod.rs
@@ -1,695 +1,32 @@
-pub mod number;
-pub mod object;
-pub mod string;
-
-use std::ffi::c_void;
-use std::mem::offset_of;
-use std::ops::Deref;
-use std::ptr::addr_eq;
-
-use jsgc::Handle;
-use jsgc::HandleMut;
-use jsgc::Trace;
-use jsgc::VisitList;
-
-use crate::Runtime;
-use crate::lambda::LambdaId;
-use crate::logger;
-
-pub use object::Object;
-pub use object::ObjectFlags;
-pub use object::Property;
-pub use object::PropertyKey;
-pub use string::String;
-
 // CAUTION: This module contains types used in JIT-generated code.  Please carefully check the
 // memory layout of a type you want to change.  It's recommended to use compile-time assertions
 // that ensure the memory layout of the type.
 
-/// A data type to hold a JavaScript value.
-//
-// DO NOT CHANGE THE ORDER OF THE VARIANTS.
-// Some operations heavily rely on the order.
-#[repr(C, u8)]
-#[derive(Clone, Debug)]
-pub enum Value {
-    None = Self::KIND_NONE,
-    Undefined = Self::KIND_UNDEFINED,
-    Null = Self::KIND_NULL,
-    Boolean(bool) = Self::KIND_BOOLEAN,
-    Number(f64) = Self::KIND_NUMBER,
-    String(Handle<String>) = Self::KIND_STRING,
-    Object(HandleMut<Object>) = Self::KIND_OBJECT,
-}
-
-static_assertions::const_assert_eq!(size_of::<Value>(), 16);
-static_assertions::const_assert_eq!(align_of::<Value>(), 8);
-
-impl Value {
-    // There is no way to define const function to extract the discriminant of each variant.
-    pub(crate) const KIND_NONE: u8 = 0;
-    pub(crate) const KIND_UNDEFINED: u8 = 1;
-    pub(crate) const KIND_NULL: u8 = 2;
-    pub(crate) const KIND_BOOLEAN: u8 = 3;
-    pub(crate) const KIND_NUMBER: u8 = 4;
-    pub(crate) const KIND_STRING: u8 = 5;
-    pub(crate) const KIND_OBJECT: u8 = 6;
-
-    pub(crate) const SIZE: usize = size_of::<Self>();
-    pub(crate) const ALIGNMENT: usize = align_of::<Self>();
-    pub(crate) const HOLDER_SIZE: usize = size_of::<u64>();
-    pub(crate) const KIND_OFFSET: usize = 0;
-    pub(crate) const HOLDER_OFFSET: usize = size_of::<u64>();
-
-    pub fn is_valid(&self) -> bool {
-        !matches!(self, Self::None)
-    }
-
-    pub fn is_callable(&self) -> bool {
-        debug_assert!(self.is_valid());
-        match self {
-            Self::Object(object) => object.is_callable(),
-            _ => false,
-        }
-    }
-
-    // 7.1.18 ToObject ( argument )
-    pub fn to_object(&self) -> Result<HandleMut<Object>, Value> {
-        match self {
-            Self::Undefined | Self::Null => Err(1001.into()), // TODO: TypeError
-            Self::Boolean(_value) => unimplemented!("new Boolean(value)"),
-            Self::Number(_value) => unimplemented!("new Number(value)"),
-            Self::String(_value) => unimplemented!("new String(value)"),
-            Self::Object(value) => Ok(*value),
-            Self::None => unreachable!(),
-        }
-    }
-
-    pub fn into_result(self, status: Status) -> Result<Value, Value> {
-        logger::debug!(event = "into_result", ?status);
-        match status {
-            Status::Normal => Ok(self),
-            Status::Exception => Err(self),
-            _ => unreachable!("{status:?}"),
-        }
-    }
-
-    // 13.5.3.1 Runtime Semantics: Evaluation
-    pub fn get_typeof(&self) -> Handle<String> {
-        match self {
-            Self::None => unreachable!(),
-            Self::Undefined => const_string!("undefined"),
-            Self::Boolean(_) => const_string!("boolean"),
-            Self::Number(_) => const_string!("number"),
-            Self::String(_) => const_string!("string"),
-            Self::Object(object) if object.is_callable() => const_string!("function"),
-            Self::Null | Self::Object(_) => const_string!("object"),
-        }
-    }
-
-    pub fn dummy_object() -> Self {
-        Self::Object(HandleMut::dummy_for_testing())
-    }
-}
-
-impl From<()> for Value {
-    fn from(_: ()) -> Self {
-        Self::Undefined
-    }
-}
-
-impl From<bool> for Value {
-    fn from(value: bool) -> Self {
-        Self::Boolean(value)
-    }
-}
-
-impl From<f64> for Value {
-    fn from(value: f64) -> Self {
-        Self::Number(value)
-    }
-}
-
-impl From<i32> for Value {
-    fn from(value: i32) -> Self {
-        Self::from(value as f64)
-    }
-}
-
-impl From<u32> for Value {
-    fn from(value: u32) -> Self {
-        Self::from(value as f64)
-    }
-}
-
-impl From<HandleMut<Object>> for Value {
-    fn from(value: HandleMut<Object>) -> Self {
-        Self::Object(value)
-    }
-}
-
-impl PartialEq for Value {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Self::Undefined, Self::Undefined) => true,
-            (Self::Null, Self::Null) => true,
-            (Self::Boolean(a), Self::Boolean(b)) => a == b,
-            (Self::Number(a), Self::Number(b)) => a == b,
-            (Self::String(a), Self::String(b)) => {
-                debug_assert_eq!(
-                    std::any::type_name_of_val(a.deref()),
-                    std::any::type_name::<String>()
-                );
-                debug_assert_eq!(
-                    std::any::type_name_of_val(b.deref()),
-                    std::any::type_name::<String>()
-                );
-                a.deref() == b.deref()
-            }
-            (Self::Object(a), Self::Object(b)) => a == b,
-            _ => false,
-        }
-    }
-}
-
-impl std::fmt::Display for Value {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::None => write!(f, "none"),
-            Self::Undefined => write!(f, "undefined"),
-            Self::Null => write!(f, "null"),
-            Self::Boolean(value) => write!(f, "{value}"),
-            Self::Number(value) => write!(f, "{value}"),
-            Self::String(value) => write!(f, "{value}"),
-            Self::Object(value) => write!(f, "object({value:?})"),
-        }
-    }
-}
-
-/// A data type to represent a closure.
-#[repr(C)]
-pub struct Closure {
-    /// An address of a lambda function compiled from a JavaScript function definition.
-    ///
-    /// This filed is initially set to a runtime function that will perform the lazy compilation of
-    /// the JavaScript function and set the actual lambda function to this field.
-    //
-    // NOTE: Using Lambda<X> instead of LambdaAddr causes some problems.  For example, functions
-    // such as `std::mem::offset_of!()` and `std::mem::align_of()` does not work with generic
-    // types such as Closure<X> even though the size of Lambda<X> is always equal to the size of
-    // usize regardless of the actual type of X.
-    pub lambda: LambdaAddr,
-
-    /// The ID of `lambda`.
-    pub lambda_id: LambdaId,
-
-    /// The number of captures.
-    ///
-    /// Usually, this field does not used in the compiled function, but we add this field here for
-    /// debugging purposes.  If we need to reduce the heap memory usage and `Closure`s dominant, we
-    /// can remove this field.
-    pub num_captures: u16,
-
-    /// A variable-length list of captures used in the lambda function.
-    pub captures: [HandleMut<Capture>; 0],
-}
-
-static_assertions::const_assert_eq!(align_of::<Closure>(), 8);
-
-impl Closure {
-    pub(crate) const LAMBDA_OFFSET: usize = std::mem::offset_of!(Self, lambda);
-    pub(crate) const CAPTURES_OFFSET: usize = std::mem::offset_of!(Self, captures);
-
-    fn captures(&self) -> &[HandleMut<Capture>] {
-        let len = self.num_captures as usize;
-        let data = self.captures.as_ptr();
-        // SAFETY: `data` is a non-null pointer to an array of pointers.
-        unsafe {
-            debug_assert!(!data.is_null());
-            debug_assert!(data.is_aligned());
-            std::slice::from_raw_parts(data, len)
-        }
-    }
-}
-
-impl std::fmt::Debug for Closure {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "closure({:?}, [", self.lambda_id)?;
-        let mut captures = self.captures().iter();
-        if let Some(capture) = captures.next() {
-            write!(f, "{:?}", capture)?;
-            for capture in captures {
-                write!(f, ", {:?}", capture)?
-            }
-        }
-        write!(f, "])")
-    }
-}
-
-impl Trace for Closure {
-    fn trace(&self, visit_list: &mut VisitList) {
-        visit_list.extend(self.captures().iter().map(|capture| capture.as_addr()));
-    }
-}
-
-/// A data type to track a captured value.
-//
-// NOTE: The `target` may point to the `escaped`.  In this case, the `target` must be updated if
-// the capture is moved during GC, so that the `target` points to the `escaped` correctly.
-#[repr(C)]
-pub struct Capture {
-    /// A captured value.
-    ///
-    /// This may point to the `escaped`.
-    pub target: *mut Value,
-
-    /// Data storage for escaped value.
-    pub escaped: Value,
-}
-
-static_assertions::const_assert_eq!(size_of::<Capture>(), 24);
-static_assertions::const_assert_eq!(align_of::<Capture>(), 8);
-static_assertions::const_assert_eq!(offset_of!(Capture, escaped), 8);
-
-impl Capture {
-    pub(crate) const TARGET_OFFSET: usize = std::mem::offset_of!(Self, target);
-    pub(crate) const ESCAPED_OFFSET: usize = std::mem::offset_of!(Self, escaped);
-
-    fn is_escaped(&self) -> bool {
-        debug_assert!(!self.target.is_null());
-        addr_eq(self.target, &self.escaped)
-    }
-}
-
-impl std::fmt::Debug for Capture {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if self.is_escaped() {
-            write!(
-                f,
-                "capture(escaped: {:?}, value: {:?})",
-                self.target, self.escaped
-            )
-        } else {
-            write!(f, "capture(onstack: {:?})", self.target)
-        }
-    }
-}
-
-impl Trace for Capture {
-    fn trace(&self, visit_list: &mut VisitList) {
-        if !self.is_escaped() {
-            return;
-        }
-
-        match self.escaped {
-            Value::String(_string) => (), // TODO
-            Value::Object(object) => visit_list.push(object.as_addr()),
-            _ => (),
-        }
-    }
-}
-
-/// A data type to represent a coroutine.
-///
-/// Memory layout:
-///
-/// ```
-/// +----------------------------------------+
-/// | Coroutine                              |
-/// +----------------------------------------+
-/// | locals[Value; num_locals]              |
-/// +----------------------------------------+
-/// | scratch_buffer[u8; scratch_buffer_len] |
-/// +----------------------------------------+
-/// | capture_buffer[u8; capture_buffer_len] |
-/// +----------------------------------------+
-/// ```
-#[derive(Debug)]
-#[repr(C)]
-pub struct Coroutine {
-    /// The closure of the coroutine.
-    pub closure: HandleMut<Closure>,
-
-    /// The state of the coroutine.
-    pub state: u32,
-
-    /// The number of the local variables used in the coroutine.
-    pub num_locals: u16,
-
-    /// The current scope ID used by the scope cleanup checker.
-    pub scope_id: u16,
-
-    /// The size of the scratch buffer in bytes.
-    pub scratch_buffer_len: u16,
-
-    /// The size of the capture buffer in bytes.
-    pub capture_buffer_len: u16,
-
-    /// A variable-length list of local variables used in the coroutine.
-    ///
-    /// `Capture::target` may point to one of `locals[]`.
-    pub locals: [Value; 0],
-}
-
-static_assertions::const_assert_eq!(align_of::<Coroutine>(), 8);
-
-impl Coroutine {
-    pub(crate) const CLOSURE_OFFSET: usize = std::mem::offset_of!(Self, closure);
-    pub(crate) const STATE_OFFSET: usize = std::mem::offset_of!(Self, state);
-    pub(crate) const NUM_LOCALS_OFFSET: usize = std::mem::offset_of!(Self, num_locals);
-    pub(crate) const SCOPE_ID_OFFSET: usize = std::mem::offset_of!(Self, scope_id);
-    pub(crate) const SCRATCH_BUFFER_LEN_OFFSET: usize =
-        std::mem::offset_of!(Self, scratch_buffer_len);
-    pub(crate) const LOCALS_OFFSET: usize = std::mem::offset_of!(Self, locals);
-
-    fn locals(&self) -> &[Value] {
-        let len = self.num_locals as usize;
-        let data = self.locals.as_ptr();
-        // SAFETY: `data` is a non-null pointer to an array of pointers.
-        unsafe {
-            debug_assert!(!data.is_null());
-            debug_assert!(data.is_aligned());
-            std::slice::from_raw_parts(data, len)
-        }
-    }
-}
-
-impl Trace for Coroutine {
-    fn trace(&self, visit_list: &mut VisitList) {
-        visit_list.push(self.closure.as_addr());
-
-        for local in self.locals() {
-            match local {
-                Value::String(_string) => (), // TODO
-                Value::Object(object) => visit_list.push(object.as_addr()),
-                _ => (),
-            }
-        }
-
-        // TODO: scan the scratch buffer and the capture buffer
-    }
-}
-
-pub trait ReturnValue {
-    fn status(&self) -> Status;
-    fn value(&self) -> Value;
-}
-
-impl<T> ReturnValue for T
-where
-    T: Clone + Into<Value>,
-{
-    fn status(&self) -> Status {
-        Status::Normal
-    }
-
-    fn value(&self) -> Value {
-        self.clone().into()
-    }
-}
-
-impl<T, E> ReturnValue for Result<T, E>
-where
-    T: Clone + Into<Value>,
-    E: Clone + Into<Value>,
-{
-    fn status(&self) -> Status {
-        if self.is_ok() {
-            Status::Normal
-        } else {
-            Status::Exception
-        }
-    }
-
-    fn value(&self) -> Value {
-        match self {
-            Ok(v) => v.clone().into(),
-            Err(err) => err.clone().into(),
-        }
-    }
-}
-
-/// Lambda function.
-pub type Lambda<X> =
-    extern "C" fn(runtime: &mut Runtime<X>, context: &mut CallContext, retv: &mut Value) -> Status;
-
-impl<X> From<LambdaAddr> for Lambda<X> {
-    fn from(value: LambdaAddr) -> Self {
-        debug_assert_ne!(value.0, 0);
-        // SAFETY: `LambdaAddr` contains only an address of a lambda function and it is always
-        // convertible to `Lambda`.
-        unsafe { std::mem::transmute(value.0) }
-    }
-}
-
-// See https://www.reddit.com/r/rust/comments/ksfk4j/comment/gifzlhg/?utm_source=share&utm_medium=web3x&utm_name=web3xcss&utm_term=1&utm_content=share_button
-
-// This function generates a wrapper function for each `host_fn` at compile time.
-pub fn into_lambda<F, R, X>(host_fn: F) -> Lambda<X>
-where
-    F: Fn(&mut Runtime<X>, &[Value]) -> R + 'static,
-    R: Clone + ReturnValue,
-{
-    debug_assert_eq!(std::mem::size_of::<F>(), 0, "Function must have zero size");
-    std::mem::forget(host_fn);
-    host_fn_wrapper::<F, R, X>
-}
-
-extern "C" fn host_fn_wrapper<F, R, X>(
-    runtime: &mut Runtime<X>,
-    context: &mut CallContext,
-    retv: &mut Value,
-) -> Status
-where
-    F: Fn(&mut Runtime<X>, &[Value]) -> R + 'static,
-    R: Clone + ReturnValue,
-{
-    // SAFETY: Parent ensured that F is zero sized and we use `ManuallyDrop` to ensure
-    // it isn't dropped (even if the callback panics).
-    #[allow(clippy::uninit_assumed_init)]
-    let host_fn = unsafe { std::mem::MaybeUninit::<F>::uninit().assume_init() };
-    let args = context.args();
-    // TODO: The return value is copied twice.  That's inefficient.
-    let result = host_fn(runtime, args);
-    *retv = result.value();
-    result.status()
-}
-
-#[derive(Debug)]
-#[repr(C)]
-pub struct CallContext {
-    /// A pointer to the call environment.
-    ///
-    /// The actual type of the value varies depending on the type of the lambda function:
-    ///
-    /// * Entry functions: 0 (null pointer)
-    /// * Regular functions: the address of the `Closure`
-    /// * Coroutine functions: the address of the `Coroutine`
-    ///
-    envp: *mut c_void,
-
-    /// The `this` argument.
-    this: Value,
-
-    /// The active function object.
-    #[allow(unused)]
-    func: Option<HandleMut<Object>>,
-
-    /// A pointer to the call context of the caller.
-    #[allow(unused)]
-    caller: *const CallContext,
-
-    /// Flags.
-    flags: CallContextFlags,
-
-    /// The depth of the call.
-    depth: u16,
-
-    /// The number of the arguments.
-    argc: u16,
-
-    /// The maximum number of the arguments.
-    argc_max: u16,
-
-    /// A pointer to the arguments.
-    argv: *mut Value,
-}
-
-impl CallContext {
-    pub const SIZE: usize = std::mem::size_of::<Self>();
-    pub const ALIGNMENT: usize = std::mem::align_of::<Self>();
-    pub const ENVP_OFFSET: usize = std::mem::offset_of!(Self, envp);
-    pub const THIS_OFFSET: usize = std::mem::offset_of!(Self, this);
-    pub const FUNC_OFFSET: usize = std::mem::offset_of!(Self, func);
-    pub const CALLER_OFFSET: usize = std::mem::offset_of!(Self, caller);
-    pub const FLAGS_OFFSET: usize = std::mem::offset_of!(Self, flags);
-    pub const DEPTH_OFFSET: usize = std::mem::offset_of!(Self, depth);
-    pub const ARGC_OFFSET: usize = std::mem::offset_of!(Self, argc);
-    pub const ARGC_MAX_OFFSET: usize = std::mem::offset_of!(Self, argc_max);
-    pub const ARGV_OFFSET: usize = std::mem::offset_of!(Self, argv);
-
-    pub(crate) fn new_for_entry(args: &mut [Value]) -> Self {
-        Self {
-            envp: std::ptr::null_mut(),
-            this: Value::Undefined,
-            func: None,
-            caller: std::ptr::null(),
-            flags: CallContextFlags::empty(),
-            depth: 0,
-            argc: args.len() as u16,
-            argc_max: args.len() as u16,
-            argv: args.as_mut_ptr(),
-        }
-    }
-
-    pub(crate) fn new_for_promise(coroutine: HandleMut<Coroutine>, args: &mut [Value]) -> Self {
-        Self {
-            envp: coroutine.as_ptr() as *mut std::ffi::c_void,
-            this: Value::Undefined,
-            func: None,
-            caller: std::ptr::null(),
-            flags: CallContextFlags::empty(),
-            depth: 0,
-            argc: args.len() as u16,
-            argc_max: args.len() as u16,
-            argv: args.as_mut_ptr(),
-        }
-    }
-
-    pub(crate) fn new_child(
-        &self,
-        func: HandleMut<Object>,
-        closure: HandleMut<Closure>,
-        args: &mut [Value],
-    ) -> Self {
-        Self {
-            envp: closure.as_ptr() as *mut std::ffi::c_void,
-            this: Value::Undefined,
-            func: Some(func),
-            caller: self,
-            flags: CallContextFlags::empty(),
-            depth: self.depth + 1,
-            argc: args.len() as u16,
-            argc_max: args.len() as u16,
-            argv: args.as_mut_ptr(),
-        }
-    }
-
-    pub(crate) fn is_new(&self) -> bool {
-        self.flags.contains(CallContextFlags::NEW)
-    }
-
-    pub(crate) fn this(&self) -> &Value {
-        debug_assert!(self.this.is_valid());
-        &self.this
-    }
-
-    pub(crate) fn func(&self) -> Option<HandleMut<Object>> {
-        self.func
-    }
-
-    pub(crate) fn closure(&self) -> HandleMut<Closure> {
-        HandleMut::from_ptr(self.envp as *mut Closure)
-            .expect("must be a non-null pointer to a Closure")
-    }
-
-    #[allow(unused)]
-    pub(crate) fn coroutine(&self) -> &Coroutine {
-        // SAFETY: `envp` is always a non-null pointer to a `Coroutine`.
-        unsafe {
-            debug_assert!(!self.envp.is_null());
-            debug_assert!(self.envp.is_aligned());
-            &*(self.envp as *const Coroutine)
-        }
-    }
-
-    pub(crate) fn coroutine_mut(&mut self) -> &mut Coroutine {
-        // SAFETY: `envp` is always a non-null pointer to a `Coroutine`.
-        unsafe {
-            debug_assert!(!self.envp.is_null());
-            debug_assert!(self.envp.is_aligned());
-            &mut *(self.envp as *mut Coroutine)
-        }
-    }
-
-    pub(crate) fn args(&self) -> &[Value] {
-        // SAFETY: `argv` is always non-null and a valid pointer to an array of `Value`s.
-        unsafe {
-            debug_assert!(!self.argv.is_null());
-            debug_assert!(self.argv.is_aligned());
-            std::slice::from_raw_parts(self.argv as *const Value, self.argc as usize)
-        }
-    }
-}
-
-bitflags::bitflags! {
-    #[derive(Clone, Copy, Debug)]
-    #[repr(C)]
-    pub(crate) struct CallContextFlags: u16  {
-        const NEW = 1 << 1;
-    }
-}
-
-/// The return value type of `Lambda` function.
-#[derive(Debug)]
-#[repr(u32)]
-pub enum Status {
-    Normal,
-    Exception,
-    Suspend,
-}
-
-static_assertions::const_assert_eq!(size_of::<Status>(), 4);
-
-/// Address of a lambda function.
-#[derive(Clone, Copy, Eq, PartialEq)]
-#[repr(C)]
-pub struct LambdaAddr(usize);
-
-impl From<usize> for LambdaAddr {
-    fn from(value: usize) -> Self {
-        debug_assert_ne!(value, 0);
-        Self(value)
-    }
-}
-
-impl<X> From<Lambda<X>> for LambdaAddr {
-    fn from(value: Lambda<X>) -> Self {
-        Self(value as usize)
-    }
-}
-
-impl std::fmt::Debug for LambdaAddr {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:p}", self.0 as *const ())
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[repr(C)]
-pub struct Promise(u32);
-
-static_assertions::const_assert_eq!(size_of::<Promise>(), 4);
-static_assertions::const_assert_eq!(align_of::<Promise>(), 4);
-
-impl Promise {
-    pub const fn is_valid(&self) -> bool {
-        self.0 != 0
-    }
-
-    pub const fn as_userdata(&self) -> usize {
-        self.0 as usize
-    }
-}
-
-impl From<u32> for Promise {
-    fn from(value: u32) -> Self {
-        Self(value)
-    }
-}
-
-impl From<Promise> for u32 {
-    fn from(value: Promise) -> Self {
-        value.0
-    }
-}
+mod call_context;
+mod capture;
+mod closure;
+mod coroutine;
+mod lambda;
+pub mod number;
+pub mod object;
+mod promise;
+pub mod string;
+mod value;
+
+pub use call_context::CallContext;
+pub use call_context::CallContextFlags;
+pub use capture::Capture;
+pub use closure::Closure;
+pub use coroutine::Coroutine;
+pub use lambda::Lambda;
+pub use lambda::LambdaAddr;
+pub use lambda::ReturnValue;
+pub use lambda::Status;
+pub use lambda::into_lambda;
+pub use object::Object;
+pub use object::ObjectFlags;
+pub use object::Property;
+pub use object::PropertyKey;
+pub use promise::Promise;
+pub use string::String;
+pub use value::Value;
